показать ошибки

+++ src/utils/Api.js
@@ -1,4 +1,3 @@
-
 // Токен: 545040d2-ca8d-4af4-bb28-cd05a11607d7
 // Идентификатор группы: cohort-61
 // Адрес сервера проекта Mesto: https://mesto.nomoreparties.co.
@@ -8,7 +7,7 @@
 // GET https://nomoreparties.co/v1/cohortId/users/me
 
 class Api {
-  constructor({basePath, headers}) {
+  constructor({ basePath, headers }) {
     this._basePath = basePath;
     this._headers = headers;
   }
@@ -22,18 +21,17 @@ class Api {
 
   getServerCards() {
     return fetch(`${this._basePath}/cards`, {
-      headers: this._headers
-    })
-    .then(this._getJson);
+      headers: this._headers,
+    }).then(this._getJson);
   }
 
   postNewCard(data) {
     return fetch(`${this._basePath}/cards`, {
-      method: 'POST',
+      method: "POST",
       headers: this._headers,
       body: JSON.stringify({
         name: data.name,
-        link: data.link
+        link: data.link,
       }),
     }).then(this._getJson);
   }
@@ -41,64 +39,59 @@ class Api {
   // Получить данные о пользователе
   getCurrentUser() {
     return fetch(`${this._basePath}/users/me`, {
-      headers: this._headers
-    })
-    .then(this._getJson);
+      headers: this._headers,
+    }).then(this._getJson);
   }
 
   // Обновить данные пользователя
   updateUserInfo(data) {
     return fetch(`${this._basePath}/users/me`, {
-      method: 'PATCH',
+      method: "PATCH",
       headers: this._headers,
       body: JSON.stringify({
         name: data.name,
-        about: data.job
-      })
-    })
-    .then(this._getJson)
+        about: data.job,
+      }),
+    }).then(this._getJson);
   }
 
   // Обновить аватар
   updateAvatar(data) {
     return fetch(`${this._basePath}/users/me/avatar`, {
-      method: 'PATCH',
+      method: "PATCH",
       headers: this._headers,
       body: JSON.stringify({
-        avatar: data.avatar
-      })
-    })
-    .then(this._getJson)
+        avatar: data.avatar,
+      }),
+    }).then(this._getJson);
   }
 
   deleteCard(_id) {
     return fetch(`${this._basePath}/cards/${_id}`, {
-      method: 'DELETE',
+      method: "DELETE",
       headers: this._headers,
     }).then(this._getJson);
   }
 
-  like(_id) {
-    return fetch(`${this._basePath}/cards/${_id}/likes`, {
-      method: 'PUT',
-      headers: this._headers
-    })
-    .then(this._getJson)
-  }
-
-  dislike(_id) {
-    return fetch(`${this._basePath}/cards/${_id}/likes`, {
-      method: 'DELETE',
-      headers: this._headers
-    })
-    .then(this._getJson)
+  changeLikeCardStatus(_id, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._basePath}/cards/${_id}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      }).then(this._getJson);
+    } else {
+      return fetch(`${this._basePath}/cards/${_id}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      }).then(this._getJson);
+    }
   }
 }
 
 export const api = new Api({
-  basePath: 'https://mesto.nomoreparties.co/v1/cohort-61',
+  basePath: "https://mesto.nomoreparties.co/v1/cohort-61",
   headers: {
-    authorization: '545040d2-ca8d-4af4-bb28-cd05a11607d7',
-    'Content-Type': 'application/json'
+    authorization: "545040d2-ca8d-4af4-bb28-cd05a11607d7",
+    "Content-Type": "application/json",
   }
 });
